requirements.txt
jupyter-book
matplotlib
numpy
PySastrawi